Add --with-ipv4 and --with-ipv6 options to cnfigure
Currently, to get IPv6 fping you have to pass -DIPV6 to to make while
compiling, and the resulting fping will not handle ipv4.  This patch
adds --with-ipv4 (enabled by default) and --with-ipv6 (disabled by
default) that gives a IPv4 capable fping, and a IPv6 capable fping6
respectively.  Both can be enabled, or either one.  If both the IPv4 and
the IPv6 versions are disabled, configure will throw an error.