Add fping6.8, manual for fping6.
This manual is generated from the same fping.pod as the fping.8 manual.
Sed is then used to s/fping/fping6/ in fping6.8.  This manual is only
installed if fping is configured with --enable-ipv6.  At the same time,
change so that fping.8 (IPv4 manual) is only installed if fping is
configured with --enable-ipv4 (default).
Add a check to configure for a working sed as well, since that's used
here.