refactor IPv4 address generation and limit CIDR
Both functions add_cidr() and add_range() end with the same code.
Extract this code into a new function add_addr_range_ipv4() and
use it.  Additionally, move the check to limit the number of
generated addresses to a compile time constant (MAX_GENERATE+1)
into the new function.

This is a change in "fping -g" behavior, because now both forms
of invoking the address generator use the MAX_GENERATE+1 limit.
Before, only "fping -g START END" would use this limit, but
"fping -g CIDR" would not.

Add a test to check if the MAX_GENERATE+1 limit is applied when
using "fping -g CIDR".