Allow error messages when sending to fe80:0:0:1::
Linux installs fe80::/64 routes when an interface has an IPv6
link-local address.  It does not install a route for the whole
IPv6 link-local prefix fe80::/10.  Thus on systems without an
IPv6 default route, sending a packet to fe80:0:0:1::, which is
outside of fe80::/64, results in an error message.

Since this test is intended to verify generator operation when
crossing the 64-bit boundary inside an IPv6 address, but not
necessarily that the system can send a packet to that destination,
a non-empty `stderr` can be accepted.

Additionally, other tests generated addresses inside fe80::/64
and those tests still require an empty `stderr`.