Address #262: Update print_per_system_splits to use stdout instead of stderr
There is inconsistency between how output is printed.
'-Q' - report_interval - prints to stderr
'-N' - netdata_flag - prints to stdout.
'-s' - stats_flags - prints to stderr

With this change, the '-Q' will also print to stdout instead of stderr.
If there is a legitimate reason why there is an inconsistency, then
please specify it in the documentation.