Catch long-only option error when using -0
The long-only options use the character '0' to select the switch
case used to handle the different long-only options.  This can be
given as a normal short option that requires an argument:

    $ fping -0
    fping: option requires an argument -- '0'
    see 'fping -h' for usage information
    $ fping -0 127.0.0.1 127.0.0.2
    127.0.0.2 is alive

Catch this usage error by requiring that one of the long-only
options is given when handling the case of '0'.