test --print-tos with an IPv6 target
Currently, --print-tos only works with IPv4.  In contrast
to IPv4, IPv6 raw sockets do not provide access to the
raw IP packet.  Thus the method used to access the TOS
byte value of an IPv4 packet does not work for the TC byte
value of an IPv6 packet.  This results in always generating
'(TOS unknown)' for IPv6 targets.